requirements.txt
playwright>=1.40
pytest>=7.0
pytest-playwright
pytest-html
allure-pytest
pytest-bdd